Make permanent firewalld changes in state.
The new behaviour is the following: All changes in the firewalld.present
state affect the runtime configuration, and at the end, the runtime
configuration overwrites the permanent configuration with

    firewall-cmd --runtime-to-permanent

This has the advantage that all firewall changes apply immediately and
survive a firewall reload. As some changes (for example --new-zone) only
apply to the permanent configuration, these changes have to be made
before all others and a reload (discarding runtime, using permanent
configuration) is triggered afterwards.

Another option would have been the other way round: Making only
permanent changes and reloading the firewall at the end of the run.